# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "CRM1 API",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/task/index": {
            "get": {
                "tags": [
                    "task"
                ],
                "summary": "Список задач",
                "operationId": "879e5cc49ca6860eb0a21ab4f3d15cf6",
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Task"
                                    }
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Error server"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/task/{id}": {
            "get": {
                "tags": [
                    "task"
                ],
                "summary": "Поиск по id",
                "operationId": "3a0ddf904fb42f5847553f140ebdff85",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "id задачи",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Task resource",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Task"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Not found"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "task"
                ],
                "summary": "Удаление по id",
                "operationId": "2cbf997202f0cefc89e1f275b995d3ca",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "id задачи",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Task deleted"
                    },
                    "404": {
                        "description": "Not found"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/task/by-user/{id}": {
            "get": {
                "tags": [
                    "task"
                ],
                "summary": "Поиск по id пользователя",
                "operationId": "3e36b27c6afc7ff3aa4a5af86a514485",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "id пользователя",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Task resource",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Task"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Not found"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/task/": {
            "post": {
                "tags": [
                    "task"
                ],
                "summary": "Создание задачи",
                "operationId": "e5f5de41f53bd8f384f176f2e6ff6770",
                "requestBody": {
                    "$ref": "#/components/requestBodies/CreateUserRequest"
                },
                "responses": {
                    "201": {
                        "description": "User resource",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Task"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Error validation"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/task/update/{id}": {
            "patch": {
                "tags": [
                    "task"
                ],
                "summary": "Обновление по id",
                "operationId": "86998812142e9a7da71ae0a69eaca07c",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "id задачи",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "requestBody": {
                    "$ref": "#/components/requestBodies/UpdateTaskRequest"
                },
                "responses": {
                    "200": {
                        "description": "User resource",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Task"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Error validation"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/task/change-dline/{id}": {
            "patch": {
                "tags": [
                    "task"
                ],
                "summary": "Смена дедлайна по id",
                "operationId": "b5c6f2511df7f4ba6e34f0d61c57087b",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "id задачи",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "requestBody": {
                    "$ref": "#/components/requestBodies/UpdateTaskRequest"
                },
                "responses": {
                    "200": {
                        "description": "User resource",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Task"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Error validation"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/task/status/{id}": {
            "patch": {
                "tags": [
                    "task"
                ],
                "summary": "Смена статуса по id",
                "operationId": "42606fe9c13d991b0b1ce8e4c07b35a4",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "id задачи",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "requestBody": {
                    "$ref": "#/components/requestBodies/UpdateTaskRequest"
                },
                "responses": {
                    "200": {
                        "description": "User resource",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Task"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Error validation"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/task/related/{id}": {
            "patch": {
                "tags": [
                    "task"
                ],
                "summary": "Смена сущности по id",
                "description": "Смета сущности по которой определена задача. client\\deal",
                "operationId": "25101ce55e788d2949a34fc5b9ee2d5b",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "id задачи",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "requestBody": {
                    "$ref": "#/components/requestBodies/UpdateTaskRequest"
                },
                "responses": {
                    "200": {
                        "description": "User resource",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Task"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Error validation"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/task/assigned/{id}": {
            "patch": {
                "tags": [
                    "task"
                ],
                "summary": "Смена ответственного по id задачи",
                "description": "Смета ответственного за задачу",
                "operationId": "fdabd4c2895917799d391afd72b37ede",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "id задачи",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "requestBody": {
                    "$ref": "#/components/requestBodies/UpdateTaskRequest"
                },
                "responses": {
                    "200": {
                        "description": "User resource",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Task"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Error validation"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "Task": {
                "required": [
                    "title",
                    "description",
                    "deadline",
                    "status"
                ],
                "properties": {
                    "title": {
                        "type": "string",
                        "example": "test title"
                    },
                    "description": {
                        "type": "string",
                        "example": "test description"
                    },
                    "deadline": {
                        "type": "string",
                        "example": "2026-02-01"
                    },
                    "status": {
                        "type": "string",
                        "example": "new"
                    }
                },
                "type": "object"
            },
...
        },
        "requestBodies": {
...
            "CreateUserRequest": {
                "required": true,
                "content": {
                    "application/json": {
                        "schema": {
                            "required": [
                                "name",
                                "email",
                                "password",
                                "is_active",
                                "role"
                            ],
                            "properties": {
                                "name": {
                                    "type": "string",
                                    "example": "Adam"
                                },
                                "email": {
                                    "type": "string",
                                    "format": "email",
                                    "example": "t@test.com"
                                },
                                "password": {
                                    "description": "Minimum 8 symbols",
                                    "type": "string",
                                    "example": "test00000"
                                },
                                "is_active": {
                                    "type": "boolean",
                                    "example": "true"
                                },
                                "role": {
                                    "description": "Из списка: admin, manager, user, viewer",
                                    "type": "string",
                                    "example": "user"
                                }
                            },
                            "type": "object"
                        }
                    }
                }
            },
...
            "UpdateTaskRequest": {
                "required": true,
                "content": {
                    "application/json": {
                        "schema": {
                            "required": [],
                            "properties": {
                                "title": {
                                    "type": "string",
                                    "example": "test title",
                                    "nullable": true
                                },
                                "description": {
                                    "type": "string",
                                    "example": "test description",
                                    "nullable": true
                                },
                                "assigned_to": {
                                    "description": "Id юзера",
                                    "type": "integer",
                                    "example": "1",
                                    "nullable": true
                                },
                                "related_type": {
                                    "description": "Из списка: client, deal",
                                    "type": "string",
                                    "example": "client",
                                    "nullable": true
                                },
                                "due_date": {
                                    "description": "Deadline",
                                    "type": "string",
                                    "example": "2026-02-01 21-22",
                                    "nullable": true
                                },
                                "status": {
                                    "description": "Из списка: new, progress, done, overdue",
                                    "type": "string",
                                    "example": "new",
                                    "nullable": true
                                }
                            },
                            "type": "object"
                        }
                    }
                }
            },
...
        },
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "bearerFormat": "Token",
                "scheme": "bearer"
            }
        }
    },
...
}